PLBD wait for connection info + message explain that user must check if he has permission when creation file fails
