+++ src/builtin_devices/panduza/test/itf_ping.rs
@@ -1,8 +1,8 @@
 
-use std::str::FromStr;
+// use std::str::FromStr;
 
 use async_trait::async_trait;
-use serde_json::{json, Value};
+use serde_json::Value;
 
 use crate::{attribute::JsonAttribute, interface::{self, AmInterface}, subscription};
 use crate::interface::Builder as InterfaceBuilder;
