+++ src/device/factory.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::vec;
 
-use serde_json::json;
+// use serde_json::json;
 
 use super::Device;
