+++ src/interface/interface.rs
@@ -1,6 +1,6 @@
 use crate::platform::services::AmServices;
 use crate::platform::FunctionResult as PlatformFunctionResult;
-use crate::platform::PlatformError;
+// use crate::platform::PlatformError;
 // use crate::platform_error_result;
 use crate::platform_error;
